chore: move syncTransaction to func
