+++ packages/element-lib/src/sidetree-v2/func.js
@@ -1,6 +1,7 @@
 /* eslint-disable arrow-body-style */
 const base64url = require('base64url');
 const crypto = require('crypto');
+const schema = require('../schema');
 
 // This function applies f, an async function, sequentially to an array of values
 // We need it because:
@@ -54,6 +55,52 @@ const batchFileToOperations = batchFile => batchFile.operations.map((op) => {
   };
 });
 
+const syncTransaction = async (sidetree, transaction) => {
+  const { transactionNumber } = transaction;
+  const anchorFile = await sidetree.storage.read(transaction.anchorFileHash);
+  if (!schema.validator.isValid(anchorFile, schema.schemas.sidetreeAnchorFile)) {
+    // TODO
+    // console.warn('anchorFile not valid', anchorFile);
+    return null;
+  }
+  const batchFile = await sidetree.storage.read(anchorFile.batchFileHash);
+  if (!schema.validator.isValid(batchFile, schema.schemas.sidetreeBatchFile)) {
+    // console.warn('batch file not valid', anchorFile);
+    return null;
+  }
+  const operations = batchFileToOperations(batchFile);
+  const operationsByDidUniqueSuffixes = operations.map((operation) => {
+    const { decodedOperationPayload } = operation;
+    const didUniqueSuffix = decodedOperationPayload.didUniqueSuffix
+      ? decodedOperationPayload.didUniqueSuffix
+      : payloadToHash(decodedOperationPayload);
+    return {
+      type: didUniqueSuffix,
+      didUniqueSuffix,
+      transactionNumber,
+      ...operation,
+    };
+  });
+  const writeOperationToCache = op => sidetree.db.write(`operation:${op.operationHash}`, op);
+  return executeSequentially(writeOperationToCache, operationsByDidUniqueSuffixes)
+    .then(() => {
+      return sidetree.db.write(`transaction:${transaction.transactionNumber}`, {
+        type: 'transaction',
+        transactionNumber: transaction.transactionNumber,
+      });
+    }).catch((error) => {
+      console.log(error);
+      // https://stackoverflow.com/questions/18391212/is-it-not-possible-to-stringify-an-error-using-json-stringify
+      const stringifiedError = JSON.stringify(error, Object.getOwnPropertyNames(error));
+      return sidetree.db.write(`transaction:${transaction.transactionNumber}`, {
+        type: 'transaction',
+        transactionNumber: transaction.transactionNumber,
+        error: stringifiedError,
+      });
+    });
+};
+
+
 module.exports = {
   executeSequentially,
   encodeJson,
@@ -61,4 +108,5 @@ module.exports = {
   payloadToHash,
   getDidUniqueSuffix,
   batchFileToOperations,
+  syncTransaction,
 };
